requirements.txt
Flask
Flask-Cors
Flask-Bcrypt
Flask-SQLAlchemy
gunicorn
PyJWT
clarifai